wifi: Update deny list entry size to uint16_t
As per the connectivity document deny list entry size should be uint16
559910_Intel_Connectivity_Platforms_BIOS_Guidelines_Rev6_4.pdf

Fixes: cc50770cd0("wifi: Add support for wifi time average SAR config")
Reviewed-on: https://review.coreboot.org/c/coreboot/+/57407
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/include/sar.h
@@ -44,7 +44,7 @@ struct avg_profile {
 	uint8_t revision;
 	uint8_t tas_selection;
 	uint8_t tas_list_size;
-	uint8_t deny_list_entry[MAX_DENYLIST_ENTRY];
+	uint16_t deny_list_entry[MAX_DENYLIST_ENTRY];
 } __packed;
 
 struct dsm_profile {
